project/backend/requirements.txt
fastapi==0.113.0
uvicorn[standard]==0.24.0
pydantic==2.11.7
pydantic-settings==2.1.0
email-validator==2.1.0
httpx==0.28.1
stripe==7.0.0
python-dotenv==1.0.0
supabase==2.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-mock==3.11.1

